rename householdsVisitPage to bulkHouseholdVisitsPage to avoid confusion
